r: `collate_fn` is created in catalog f: implement dataloader for mmdetection
