+++ configs/vision/object-detection/mmdet/reference/faster_rcnn_r50_caffe_fpn_mstrain_1x_coco.yaml
